pyproject.toml
[tool.poetry]
name = "duga"
version = "0.1.0"
description = ""
authors = ["Fedor Sokolov <itsfedorsokolov.gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.11.9"
fastapi = "0.115.4"
uvicorn = "^0.32.0"
pydantic = {extras = ["email"], version = "^2.10.6"}
pydantic-settings = "^2.6.0"
python-dotenv = "^1.0.1"
sqlalchemy = "^2.0.37"
alembic = "^1.15.1"
asyncpg = "^0.30.0"
greenlet = "^3.1.1"
python-jose = "^3.4.0"
bcrypt = "4.0.1"
boto3 = "1.35.99"
python-multipart = "^0.0.20"
jinja2 = "^3.1.6"
aiofiles = "^24.1.0"
redis = "^6.1.0"



[tool.poetry.group.dev.dependencies]
ruff = "0.11.2"
notebook = "^7.2.0"
pre-commit = "^3.7.1"
pytest = "^8.3.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.ruff]
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv"
]
# Include jupyter checks
extend-include = ["*.ipynb"]

line-length = 120
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",
    "F",
    "I",
    "N",
    "UP",
    "ANN",
    "RUF",
    "ASYNC"
]
ignore = ["ANN101", "RUF001", "RUF002", "RUF003"]

# Allow fix for all enabled rules (when `--fix`) is provided.
fixable = ["ALL"]
unfixable = []

[tool.ruff.format]
# Like Black, use double quotes for strings.
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
